# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Estimation Platform API
  description: Comprehensive API for project management, task estimation, and AI-powered assistance
  version: 1.0.0
  contact:
    name: API Support
    email: support@ai-estimation.com
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.ai-estimation.com
    description: Production server
security:
  - sessionAuth: []
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie
  schemas:
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        objective:
          type: string
        status:
          type: string
          enum: [ACTIVE, ARCHIVED]
        githubUrl:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /api/projects/{id}:
    get:
      summary: Get project by ID
      tags:
        - Projects
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'

    put:
      summary: Update project
      tags:
        - Projects
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                objective:
                  type: string
                status:
                  type: string
      responses:
        '200':
          description: Project updated

    delete:
      summary: Delete project
      tags:
        - Projects
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project deleted
...